add  bankOperationRepository  BankOperationService and her implementation

+++ src/main/java/com/helscorp/banking/serviceImpl/UserServiceImpl.java
@@ -56,6 +56,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void delete(Integer id) {
+        if(id == null){
+            return;
+        }
         repository.deleteById(id);
     }
